feat: upload app.js
Upload app.js- the file where most of the work for the project is completed. Current functionality allows game board, characters, and enemies to be rendered. Enemies move across the screen and the character movement can be controlled using the arrow keys. The game is won when the player reaches the water and resets.